% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparsers.R
\name{args}
\alias{args}
\title{Parse multiple arguments}
\usage{
args(ARGV, parser, ...)
}
\arguments{
\item{ARGV}{(All functions) The output of \code{\link[=getKVArgs]{getKVArgs()}}}

\item{parser}{One of the \verb{arg_*} argument parsers}

\item{...}{Arguments forwarded to the function specified as \verb{parser=}.}

\item{matching}{A regular expression. Process only arguments matching this.}

\item{exlude}{A regular expression. Do not process arguments matching this.
This has precedence over \verb{matching=}.}
}
\description{
Parse all matching arguments in \code{ARGV} using any \verb{arg_*} parsing function.
}
\examples{
ARGS <- getKVArgs(c("A=1", "B:X=10", "B:Y = 11"))
args(ARGS$B, arg_int)

\dontrun{
ARGS <- getKVArgs(c("A=1", "B:X=aaa", "B:Y = 11"))
args(ARGS$B, arg_int)
## Error: Cannot parse integer argument B:X: NAs introduced by coercion
}
}
