% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSumm.R
\name{dataSummary}
\alias{dataSummary}
\title{Summarise data columns}
\usage{
dataSummary(d, file = "", signif = 3, ...)
}
\arguments{
\item{d}{A data.frame to print a summary of}

\item{con}{Connection to print to. Forwarded to \code{\link[base:cat]{base::cat()}}.}
}
\description{
For each column of the data.frame \code{d}, print a line of summary. Elements and
their counts are summarized for characters and factors, for numeric values,
minimum, quartile 1, median, quartile 3, maximum, mean and number of NAs are
printed, in that order. For matrices, the dimensions are shown. Data frames
are summarized using a shorter form of the above descriptions for each
column.
}
\examples{
dataSummary(diamonds)
## 53940 x 10 tbl_df, tbl, data.frame:
## $ carat   <dbl>:    0.200 -[    0.400     0.700      1.04 ]-     5.01 µ=    0.798 NAs: 0
## $ cut     <fct>: Ideal: 21551 Premium: 13791 Very Good: 12082 Good: 4906 Fair: 1610  NA: 0
## $ color   <fct>: G: 11292 E: 9797 F: 9542 H: 8304 D: 6775 I: 5422 J: 2808  NA: 0
## $ clarity <fct>: SI1: 13065 VS2: 12258 SI2: 9194 VS1: 8171 VVS2: 5066 VVS1: 3655 IF: 1790 I1: 741  NA: 0
## $ depth   <dbl>:     43.0 -[     61.0      61.8      62.5 ]-     79.0 µ=     61.7 NAs: 0
## $ table   <dbl>:     43.0 -[     56.0      57.0      59.0 ]-     95.0 µ=     57.5 NAs: 0
## $ price   <int>:     326. -[     950.  2.40e+03  5.32e+03 ]- 1.88e+04 µ= 3.93e+03 NAs: 0
## $ x       <dbl>:     0.00 -[     4.71      5.70      6.54 ]-     10.7 µ=     5.73 NAs: 0
## $ y       <dbl>:     0.00 -[     4.72      5.71      6.54 ]-     58.9 µ=     5.73 NAs: 0
## $ z       <dbl>:     0.00 -[     2.91      3.53      4.04 ]-     31.8 µ=     3.54 NAs: 0
}
