% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKVArg.R
\name{getKVArgs}
\alias{getKVArgs}
\title{Parse the command line containing key=value pairs}
\usage{
getKVArgs(args = commandArgs(TRUE))
}
\description{
Convert a command line (character vector) into a list of key/value pairs.
}
\examples{

getKVArgs(c("A=1", "B=2"))
## $A
## [1] "1"
##
## $B
## [1] "2"

# Keys with colons lead to a list argument:
getKVArgs(c("A:X=1", "A:Y = 2", "B=2"))
## $A
## $A$X
## [1] "1"
##
## $A$Y
## [1] "2"
##
##
## $B
## [1] "2"

# Use "+=" to create argument vectors
getKVArgs(c("A+=1", "A+=3", "B=2"))
## $A
## [1] "1" "3"
##
## $B
## [1] "2"

\dontrun{
getKVArgs(c("A:X=1", "A:X:Y=3", "B=2"))

## Error in setL(x[[path[1]]], path[-1], val, c(basePath)) :
##   Element A:X:Y -- cannot mix list and non-list elements

getKVArgs(c("A+=1", "A:X=3", "B=2"))
}

getKVArgs(c("A:B:C=1", "A:B?:test=hello", "A:B:D=3"))
## $A
## $A$B
## [1] 1
## -attr(, "test")
## [1] "hello"

}
