% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_io.R
\name{load_image}
\alias{load_image}
\title{Load an environment}
\usage{
load_image(filename, envir = .GlobalEnv, ...)
}
\arguments{
\item{filename}{File to read}

\item{envir}{Environment to write into}

\item{...}{Arguments to \code{qs::qread()}}
}
\description{
Load an environment
}
