% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRDS_mt.R
\name{save.image}
\alias{save.image}
\title{Save the workspace using multithreaded compression}
\usage{
save.image(file, compress = TRUE, envir = .GlobalEnv, precheck = FALSE, ...)
}
\description{
Save the workspace using multithreaded compression
}
