% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_io.R
\name{save_image}
\alias{save_image}
\title{Save the global environment}
\usage{
save_image(filename, envir = .GlobalEnv, all = FALSE, ...)
}
\arguments{
\item{filename}{File to write to}

\item{envir}{Environment to save}

\item{...}{Arguments to \code{\link[qs:qsave]{qs::qsave()}}}
}
\description{
Save the global environment
}
